requirements.txt
attrs==19.3.0
awscli==1.16.57
bcrypt==3.1.7
beautifulsoup4==4.8.1
botocore==1.12.47
CacheControl==0.12.5
cachy==0.2.0
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
cleo==0.6.8
Click==7.0
colorama==0.3.9
coverage==4.5.3
cryptography==2.8
Django==2.1.5
django-autofixture==0.12.0
django-cors-headers==2.5.2
django-crontab==0.7.1
django-extensions==1.6.1
django-filter==0.12.0
django-nose==1.4.6
django-webpack-loader==0.6.0
djangorestframework==3.8.2
djangorestframework-simplejwt==4.3.0
docker-py==1.10.6
docker-pycreds==0.4.0
docutils==0.14
Fabric==1.10.2
flake8==3.6.0
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
jmespath==0.9.3
jsonschema==3.2.0
lockfile==0.12.2
mccabe==0.6.1
msgpack==0.6.2
nose==1.3.7
numpy==1.17.3
pandas==0.25.3
paramiko==2.6.0
pastel==0.1.1
pep8==1.7.1
pip-tools==4.2.0
pkginfo==1.5.0.1
poetry==0.12.17
psycopg2==2.8.4
pyasn1==0.4.4
pycodestyle==2.4.0
pycparser==2.19
pyenv-win==1.2.4
pyflakes==2.0.0
PyJWT==1.7.1
pylev==1.3.0
PyNaCl==1.3.0
pyparsing==2.4.5
pyrsistent==0.14.11
python-dateutil==2.7.5
pytz==2018.5
PyYAML==3.13
requests==2.22.0
requests-toolbelt==0.8.0
retrying==1.3.3
rsa==3.4.2
s3transfer==0.1.13
scrapinghub==2.2.1
selenium==3.141.0
shellingham==1.3.1
shub==2.10.0
six==1.11.0
soupsieve==1.9.5
sqlparse==0.3.0
toml==0.10.0
tomlkit==0.5.8
tqdm==4.37.0
urllib3==1.24.1
virtualenv==16.7.7
virtualenvwrapper-win==1.2.5
webencodings==0.5.1
websocket-client==0.56.0
xlwt==1.3.0
